refactor: unify mongoose singleton and debug
